refactor: move middleware into api
