+++ internal/api/routes/health_check.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/guuzaa/email-newsletter/internal/middleware"
+	"github.com/guuzaa/email-newsletter/internal/api/middleware"
 )
 
 func healthCheck(c *gin.Context) {
