+++ internal/api/routes/subscriptions_confirm.go
@@ -4,9 +4,9 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/guuzaa/email-newsletter/internal/api/middleware"
 	"github.com/guuzaa/email-newsletter/internal/database/models"
 	"github.com/guuzaa/email-newsletter/internal/domain"
-	"github.com/guuzaa/email-newsletter/internal/middleware"
 	"gorm.io/gorm"
 )
